Separate class for recurrence data
Implement a separate class for recurrence data and embed it into the
CalendarEvent class.  This encapsulates the recurrence data and makes
it cleaner to implement the interface "MessageSourceResolvable".

Further implement "MessageSourceResolvable" to generate a localized
string from the recurrence data.  This is used in show view and later
in list view.