Calendar event reminders
Implement storage and retrieval of reminders for calendar events in the
form of the create and edit view. Reminders are stored on user base
with a default set of reminders which acts as fallback settings if a
user has not defined own reminders.

I add the property "owner" to the CalendarEvent class to handle event
owners correctly.

I change the behavior in the FixedSelAutocomplete class in JavaScript.
Now it can obtain the fetch URL to retrieve the autocomlete values from
the HTML attribute "data-find-url".

Furthermore, I source out the JavaScript code in _form.gsp and create a
new JavaScript file.