Widget for extended autocompletion
Implement a jQuery widget for extended autocompletion. The widget
encapsulates all code which was in the former FixedSelAutocomplete
class. It provides an autocomplete input control which may optionally
be displayed as combobox. The control forces the user to select an item
from the list. If so, it stores the ID of the item in an associated
hidden field.