Revised controllers according to Grails 2.0
First I removed the logic that was necessary in Grails 1.x to handle the
invoicing items which belong to an invoicing transaction. Now it is
possible to remove all null items from the submitted list and to save
the list.

Furthermore, I change the if/then login in the controllers for better
readability.

After all, I fix a bug in the letter bar tag.