Bugfix #212
A non-admin user could not change the stage of an invoice because he
could not see the edit button any more. This is valid for dunnings and
credit memos, too.

I implement two new actions "editPayment" and "updatePayment" together
with the associated view to display a form containing only the stage
selector and the payment fields. The stage selector displays the values
greater than or equals to "delivered" only. In the show and list views
I add a button to edit the payment information.