System settings, part VI
I finish implementation of the selector values editor. I add the still
missing selector lists, prevent editing of disabled items, and solved
submission of disabled items to store the order of the items correctly.

As last, I add a container around the selector lists to clip its content
and made it scrollable.