SpringCRM installer
I implement a controller which acts as installer for SpringCRM.  The
installer offers the following features:

- create an administrator user account
- import a set of localized base data such as selector list values,
  currencies, and system settings
- a welcome and a finish page
- securing the installer by expecting the file ENABLE_INSTALLER in
  web-app/WEB-INF/data/install

The base data are loaded from a special crafted SQL file.  I also add a
sed script to convert the database initialization SQL file in
scripts/db/init-db.sql to such a crafted SQL file.

At last, the message resources are restructured because I use the TapiJi
plugin for Eclipse to ease translation.