I18n issues
Anew, I revise the i18n message generation for JavaScript.  The last
approach was erroneous because the i18n.js file was generated once for a
particular language, only.  If another user with another language would
access the application the file was not generated anew but remains in
the original language.

Now the JavaScript message resources are generated dynamically and I
remove the i18n resource mapper.