Generic template for printing
First, I implement a generic system for print templates (also see
bug #209).  Now, there is a system folder and a user folder which may
contain an arbitrary number of templates.  Each template is contained in
a directory in either the system or the user folder.  The system folder
is located in WEB-INF/data/print, the user folder in ~/.springcrm/print.
The default template in the system folder is called "default".

Any template folder with the same name in the user folder overwrites the
corresponding folder in the system folder.  Doing so, a user can
overwrite the system templates.

Each template folder may contain a file "meta.properties" which is a
standard Java properties file.  That file may contain a key "name" which
represents the name of the template for display in the GUI.
Furthermore, mainly for system templates, the name may also be stored in
the i18n files under the name "print.template.**name**", where **name**
is the directory name of the template folder.

I add a popup menu to all print buttons in the application which
displays the templates obtained from the system and the user folder as
described above.

Second, I implement a new page in the installer where the user can
define data of its company (name, address, phone, e-mail, bank data
etc.).  The data are used in print templates (see below).

At last, I create a generic template for printing invoices, quotes, etc.
based on the AMC World Technologies GmbH template.  To make the
templates generic I use client data which were set in the installer.
The old AMC World Technologies GmbH templates are now moved to the user
folder.