Bugfixes #218 and #219
Bug #218: an NPE occurred when copying purchase invoices.  I fix this
error by also adding "save" and "update" to ViewFilters.

Bug #219: an error occurred when adding new invoicing items to an
invoicing transaction and at least one of the new items was move up- or
downwards.  I fix this error by restoring the work-around I implemented
before migrating to Grails 2.0.0.  It seems that the error appears anew
in Grails 2.0.1.

For convenience, I add a SQL script to find invoicing items with a gap.