Standalone application
Mainly I make SpringCRM ready for standalone usage.  I create a new
environment "standalone" which uses an H2 database and install plugin
"standalone" to generate a JAR containing Tomcat, compiler, and this
web application.

Because I tested the application with Apache Derby I notices that some
table and column names are not conform to the SQL standard.  So I fix
these names and generate a database changelog.

Furthermore, I noticed that method ServletContext.getRealPath() does not
work when the application is loaded from a WAR file.  Then this method
returns null.  I fix it by obtaining the input streams or retrieving the
contents of folders by ServletContext.getResourcePaths().

At last, I add some checks to the XSL-FO templates to only render the
bank data if they have been set in the system settings.