Bugfix #222
Redirects to invalid URLs.  If the return URL was used when redirecting
after saving or cancelling a save/update operation an invalid URL was
generated causing the request of a non-existing page.

I fix this by adding attribute "absolute" to the createLink tag in order
to generate absolute URLs which can be used in redirect() without
problems.

At last, I activated the automatic database migration in live
environment.