Bugfix #223
Client data were not saved in installer.  I fix the bug by changing the
order of pages in the installer.  Now, the page to install base data
comes before storing the client data and creating the administrator
account.

Furthermore, I fix a bug in DataSource.groovy because a GString was not
enclosed in double quote marks.