Bugfix #224
RTE in notes form not displayed.  The bug occurred because the
_input.gsp template for the content field in the note form was not
executed.  The reason why it was not executed lies in
views/_fields/default/_field.gsp where the <textarea> control is
directly rendered instead of calling the <f:input /> tag.  However, we
actually cannot use <f:input /> because it doesn't render all other text
areas correctly.  In fact, they are rendered as <input type="text" />
tags.