Bugfix #217
Include application version in error report.  I add a Gant script which
executes each time the source is compiled.  This script generates a
sequential build number, the build date, and the build profile and
stores it in application.properties.  In the error.gsp file I add a new
block to the generated XML error report which renders these data.

In order to organize the Gant scripts correctly I move the previous
files and directories from the script folder to a new resource folder.

At last, I create a XML Schema file to validate the error report XML.