Bugfix #215
LRU entries do not show new name after renaming.  I add an additional
filter after calling update in a controller which registers the content
item with the new name.  For it I also change the LruService class which
did not update the name when storing an already existing LRU entry.

Furthermore, I noticed that the save() action of the controllers does
not create an LRU entry if they don't to the redirect show view.  I fix
this bug by storing the ID of the created content item and then applying
the filter as at the update action.

At last, I add headers and comments to the modified controllers and
domain classes.