Database migration fix
Mainly I create a fresh base changelog for database migration.  I
recognized that either using the dbCreate feature of GORM and some
single changelogs for the database-migration plugin are not enough.
When I set dbCreate to "none" not even the base structure is created.
So I add a base changelog which produces the current domain models
database structure.  Further versions of this application must produce
changesets and apply them on application boot.

Furthermore, I show additional information on the system information
page, such as pathes and the name of the executing user.

As last, I fix a style bug in the headers of the overview panels.