Security fix for installer
Mainly, I fix a security problem in the installer.  Former, the
installer was enabled during the boot time if the install status was
zero or undefined.  However, if the user called the web site much later
the installer was enabled which could be exploited by an attacker.  Now,
the installer is activated just when the user calls the web site, i. e.
in the login action of UserController.

Also I fix a bug in the base data SQL for Austria which contained a
duplicate record ID.