Generalize storage pathes
I make all storages pathes for the application configurable and general.
The pathes may be configured in Config.groovy.  The base folder for all
storage pathes may also be set by an environment variable or a Java
property.  This is necessary to get the application working on servers
like CloudFoundry which do not support a home directory.

At last, I make some minor changes on the system info page.