Bugfix #230
No labels in overview page panels in CloudFoundry.  The bug was caused
in method OverviewPanel.getTitle().  This method expects a locale to
translate the title.  But no caller submitted the locale.  getTitle()
then felt back to obtain the default locale which is not the same as the
locale requested by the client which is stored in LocaleContextHolder.
I have to submit the locale from LocaleContextHolder each time the
method is called.