Bugfixes for CloudFoundry
First, I fix the bug that the location of the searchable index was
wrong.  The computed path contained illegal parts which was caused by
the baseDir definition in Searchable.groovy.  Such a definition is not a
string but a Searchable configuration object.

Then I fix the bug that the injected variable grailsApplication in
OverviewPanel was null.  Now I access the applicationContext via
ServletContextHolder.