Bugfixes #232 and #238
Bug #232: Quotes are created twice.  This was not an error in SpringCRM
but a problem of the AJP connector.  It seems the error doesn't occur
any longer.

Bug #238: NPE when saving invoice.  I fix this bug by adding a condition
in the LruFilters class which checks if the save action actually
provides a model.  In the case of a user input error "save" redirects
directly to the create action without providing a model.  Then the LRU
filter relies on the model which is null.

Furthermore, I make some minor changes in gsp files and fix a bug that
non-admin users could not access the user settings pages.

Also, I upgrade to Grails 2.0.3.