Display unpaid amount
I add a colored text after the payment amount field which shows the
still unpaid amount.  The text is updated either when the payment amount
or the total price changes.  Depending on the sign of the still unpaid
amount the text is colored.  Also, I add a link to add all still unpaid
amount to the payment amount field.