Project items (part II)
I extend the project items list in the project show view.  I group the
items by project phase and controller and add icons for each type of
item.

I also add filters which update or delete a project item when the
associated content element is updated or deleted, respectively.

Furthermore, I remove the test for delete confirmation from all
controllers and move it to the ViewFilters class.