Bugfix #240
Error after converting quote to invoice.  I fix the bug by changing the
way a row is removed. The input controls of all rows after the deleted
one are re-numbered so that no gap in numbering occurs. Also, in the
_onSubmit() function the _removeRow() function is called to remove an
empty row. This guarantees the input controls of all subsequent rows are
re-numbered, too.