Project items (part V)
I implement a selector for the project status in the project show view.
The user can quickly select the status of the project which is also
displayed in the right margin.  Additionally, I color the status in the
list views.  To implement the selector I use the jQuery plugin
"selectmenu".

Furthermore, I add the recipient person name to all print versions of
invoicing transactions.