Fix balancing issues
I change the way balances are computed in sales journal.  Now credit
memos are taken in account to compute the balance of an invoice or
dunning.  Credit memos no longer have to be paid and any payment is not
transferred to the associated invoice or dunning.  Instead I implement
the properties "balance" and "closingBalance" for invoices, dunnings,
and credit memos.

Furthermore, I change the way the color indicators work.  They now
consider the closing balance instead of the balance.  Also, I add a
color coding for the balance values itself.