More balancing issues
I further change the way balances are computed and displayed.  In this
commit I rewrite the code to display the unpaid amount in the invoice,
dunning, credit memo, and purchase invoice forms.  I also rework the
files "invoicing-transaction-form.js" and "purchase-invoice-form.js"
which now acts as a jQuery widget instead of a loose couple of
functions.

Second, I add the display of balances to the list views of invoices,
dunnings, credit memos, and purchase invoices including coloring of the
stage and the balance.

Third, I change the way the data attributes are read from JavaScript.
Now I use the data () function of jQuery instead of the attr() function.

At last, I change some styles and add the "new purchase invoice" entry
to the quick menu.