Use classes for content tables
I add CSS classes to each column of content tables which provide the
following information:

  - the type of data which is represented in the column.  Currently, I
    defined the following types: id (for numbers and IDs), ref (for
    references to other content items), number (for decimal numbers),
    currency (for currency values), status (for status and stage
    values), and string (for all other values).
  - a unique CSS class in the form "content-table-column-domain-prop"
    where "domain" is the name of the domain name in lower cased, dashed
    form and "prop" is the name of the property, also in lower cased,
    dashed form.

I also add IDs to the column headers and refer to it in the body cells.
The headers are named "content-table-headers-domain-prop" with the same
substitution values for "domain" and "prop" as described above.

Then I styled the columns in styles.css.  Mostly, I add styles for
column data types such as dates, numbers, currencies, and IDs.

At last, I move the menu item "Purchase invoices" to the end of the menu
"Assets".