Bugfix #247
Cannot authenticate at Google.  I fix the bug by re-activating the
previously implemented OAuth 2.0 technique and make it ready for
synchronizing contacts.  The service classes now use OAuth 2.0 which
must be pre-configured by the user.  Then all Google access attempts can
be unattended where the service class also refreshes the access tokens
automatically.

In order to use the new Google API I upgrade the library files.  It was
necessary to downgrade the guava library to r07 because of a bug in the
Google library implementation.

Furthermore, I move all error views to folder "grails-app/views/errors"
and add a view which is displayed if the user doesn't have access to
Google.

At last, I add the about controller to the exclusing list in the
LoginFilters class.