Fix synchronization issues with Google
I fix a problem when synchronizing lots of contacts with Google.  After
an initial synchronization many local entries were deleted in the second
synchronization step.  The reason was that Google does transfer only a
particular number of items in the retrieve phase.

Furthermore, I fix a bug in a database migration file and remove the
Google sync button in the person show view.