Implement job to synchronize automatically
I implement a job to synchronize person entries with Google unattended.
The job is executed via Quartz with an interval defined by the
administrator.  I order to use the existing synchronization code I have
to move all this code to a separate class in order to instantiate one
object per user.  However, in the new classes I need access to the
GoogleOAuthService class so that I have to use dependency injection.

At last, I rename some files to emphasise the synchronization feature.