Bugfix #253
Items vanish when sorting in embedded list views.  The reason for the
bug was that the ID of the superior record was not submitted in the
query beyond the inital query.  For instance, say you show an
organization.  Then it also shows all invoices of this organization.
During the initial load of invoices the organization ID was submitted.
However, if you sort a column, change the page etc. the organization ID
was no longer submitted resulting in an empty embedded list.

I divide the data-load-url attribute into a URL calling the action and a
second attribute data-load-params which contains a query string fragment
which is submitted with each request.  To do it conveniently I add the
jQuery BBQ (Back Button & Query Library) to the JavaScript resources.