Bugfix #249
Cannot show or edit user which is currently logged in.  This bug was
caused because the user object from the session was used to store
relations.  In this case an LRU entry should be stored, but the
associated user object doesn't come from the database but from the
session (session.user).  I change this behavior and obtain the user from
the database.  If at other places a NonUniqueObjectException will be
thrown the problem can be fixed the same way.

I also fix a problem that an empty module list string is splitted as a
one element list.  Furthermore, I hide the display of the allowed
modules in the show view if the user is an administrator or the list is
empty.