Bugfixes #211 and #255
Bug #211: Browsing in jQuery dialog does not work in Google Chrome
(re-opened).  The bug was caused by expression
"$form.get(0).search.value" where "search" was null.  I don't know
exactly why this is the case in Chrome but I rewrite it to
"$form.find("[name=search]").val()".

Bug #255: Cannot create invoicing transactions because of form errors.
The bug was caused by a missing semicolon in the code.  Because the
Groovy compiler doesn't require semicolons this line and the following
instance initializer were interpreted as closure.