Change error and warning handling
First, I create the class ConnectorThrowable as parent for the classes
ConnectorException and ConnectorWarning to avoid code duplication.

Then, I change the constructor for ConnectorException and
ConnectorWarning which now allow an arbitrary amount of arguments of
the types String, ConnectorError, and ConnectorThrowable.  Using these
constructors we can now return error codes and associated parameters to
the client.  I also rework all classes where ConnectorException and
ConnectorWarning are thrown to use parameters if necessary.

At last, I change the return value of Volume.dir() and Volume.file()
which now return null if the directory or file, respectively, do not
exist.