Implement documents by organization
I add an embedded list to the organization show view to display the
documents of that organization.  The documents for an organization are
retrieved from the document folder using a path specification defined by
the administrator.  That specification may contain the %o placeholder
which is substituted by the organization name.

I also add a new field to the organization database table to store
(cache) the %o placeholder substitution value.