Finish documents by organization
I finish implementing the list of documents of an organization.  I
change the way the documents are referred by using the methods of the
volume classes.  So I don't need to submit the organization ID when
downloading or deleting a document.  Also, the cache problem is fixed
because now I use the methods of Volume to download or delete a
document.

Furthermore, I add a system configuration option to disable the cache in
the volume classes.