Implement document selector in projects
I add the document type to the item selector in the project's show view.
If selected, it displays an ElFinder file manager which adds the
selected file to the project.  For it, I create a new domain model class
ProjectDocument.

Because we need the local filesystem volume at various places I move the
code to obtain it to the FileService class.

At last, I update the l18n files of ElFinder.