Complete client-side computation of pricing
I finish working on the client-side computation of pricing items.  I
implement that unit prices and total prices of pricing items and the
total price of the whole pricing table is computed.  Furthermore, I
implement that related items can be selected using a target button.

I also add method reverse() to jQuery to reverse the content of a jQuery
object.