Bugfixes (amongst others #257)
First, I fix an undefined symbol in the GoogleSync class which was not
imported.

Then, I fix a bug that user defined configuration files were not loaded.

At last, I check bug #257 because the document selector was not
displayed.  But this problem was caused by some invalid data in the
browser cache.