Rework CSS class names
I change the pricing form and reduce the number of steps to calculate
the sales price.

During work I recognize that invoicing item elements and pricing
elements share a lot of classes and design.  So I rework the CSS classes
in the following ways:

- use relative class names instead of absolute: use ".price-table
  .total-price" instead of ".pricing-items-total-price"
- assign class names by a "why?" question: why the text should be
  printed bold? Because it's a total price.

By the way, I also remove <a> tags with an image inside which are used
by JavaScript only, so only the <img> tag remains. This prevents use of
misleading <a href="#"> tag or worse: <a href="javascript:void(0);"
tags.