Another try to fix user settings bug
I attempt to fix the bug when storing user settings another way.  The
bug produces the following errors:

- entry in user_data_settings table already exist
- corrupt session object

Now I no longer modify the user object in the session and in the
database separately but use the session user object only and attach it
to the persistence session before saving.  Maybe this fixes the bug.

I also convert the project for use with Groovy & Grails Tool Suite.