Final working version with plugins
I finally fix the problem that cloud-foundry plugin is necessary for
run-app but its JAR files should not be included in the resulting WAR
file.

Furthermore, I move jsunit to the test scope and svn to build in order
to remove it from the resulting WAR, too.  Now, the WAR file is about 10
MB smaller than before.