Bugfix #261
First I fix bug #261: NPE when saving a purchase invoice.  It was caused
by a missing constraint for field "tax" and a missing check for null
values when computing the tax classes.  Nonetheless, the null check is
necessary because the tax classes are computed before validation.

After fixing this I recognized that the user doesn't get information
about what is wrong in items of invoicing transactions.  So I add a
service class and a method to obtain the error messages for embedded
items.