User settings as map
Mainly, I implement the user settings as a Map implementation.  The map
supports inserting, updating, and deleting settings when it is changed.
This is a far better approach than the user settings in the last commit.

Furthermore, I fix a bug in scripts.js that the currently selected font
size is not marked in the view.