Use class Currency for currencies
I rework the code which uses currencies (mainly ViewTagLib) to use the
class Currency and a three-letter currency code such as EUR, USD etc.
instead of the currency symbol.  This has the advantage that we can
obtain the default number of fraction digits for a given currency and
can use the attribute currencyCode instead of currencySymbol, where the
former is preferred.

I don't implement code to convert all currency symbol settings to the
new currency code.  I implement, however, code that falls back to EUR if
a currency symbol is used.  Users running this application should reset
their currency in the system settings menu.