Configurable number of fraction digits
I make the number of fraction digits for all currency values
configurable.  For it, I change the types of the currency columns to
"decimal(29, 10)".  Then, I rewrite the ViewTagLib and the JavaScript
library scripts.js which now format all currency values with the
configured number of fraction digits.  In order to get the JavaScript
working the number of fraction digits is rendered into the <html>
element where the number is written to the model by the ViewFilters
class.  Also, the currency symbol and the locale is written to the
<html> tag.