Refactor sales item pricing code
I move the common code for sales item pricing to the separate
salesItemPricing.gsp in grails-app/views/layouts which is included by
the product form as well as the service form.  I also optimize the
display of input errors in the pricing section and add a possibility to
remove pricing.

I also move common code to save sales items and their pricings to the
service class SalesItemService.  I fix a bug with endless dependencies
when saving domain classes by removing the foreign key in
SalesItemPricing to SalesItem.

At last, I add some unit test cases for SalesItemPricing and
SalesItemPricingItem.