Test show and list view of organization
I finish implementing the functional test case for the show view of
organizations, implement a functional test case for the list view of
organizations, and start implementing the functional test case for the
edit view of organizations.

Furthermore, I change the way of rendering links to URLs because in the
last release I removed the url constraint.  So I use the widget
parameter in the constraints now.