DbUnit integration
I add a plugin to use DbUnit in the functional tests.  The former
commits implemented tests which were valid in a particular order.  To
allow execution in random order, I needed a deterministic way to reset
the database to a defined state.  So I implement data sets for installer
test (which empties some tables) and for all other tests (which fills
some tables with installation fixture data).  In order to have the same
fixture data as there are in the base-data-*.sql files I write a script
which can be called on command line using

    grails convert-installation-fixture [locale]

to convert the base-data-[locale].sql file to a DbUnit data set.  If
"locale" is not specified, it defaults to "de-DE".

Then I move the code to redirect to the installer from UserController to
LoginFilters.  I also move the redirect to the installer from
InstallFilters to LoginFilters and remove the former class.