Empty all tables in base data
I implement a test case in InstallationTest working on a previously
completed installation.  This test reveals that we cannot replace
contents of some tables such as sel_value or seq_number during
installation because it injures referential integrity.  So we have to
empty all database tables associated to a domain model class.  However,
if we do this, we need to strongly ask for confirmation before
truncating the tables.  So I implement a emphasized warning if existing
data are about to be overwritten as well as a double confirmation when
the user clicks "Next".  I also remove that warning if the user performs
a fresh installation.

Additionally, I add the possibility for empty lines and comments to the
base-data-*.sql files and I change the convert-installation-fixture
script to reflect the changes described above.

At last, I further develop the test case for existing organizations.