Implement person functional test case
I implement a function test case to test the person controller and
domain model.  The test revealed an error when creating a person without
picture.  Then an empty picture (0 bytes) are stored.  I fix this.

Furthermore, I call the modules (persons, organizations) in the test
cases using a URL instead of using the menus.  This is because the menus
must first open (slide down) and this makes impossible to select the
correct submenu item.