Implement credit memo functional test
I implement the functional test for credit memos and, along this way,
fix some problems in the views and the domain model class.

I also notice a potential bug in the views: the decimal separator and
the grouping separator for numbers have not been rendered in the <html>
tag as data attributes.  I add the separators to the current model in
the ViewFilters class and render them in the <html> tag in the main.gsp
layout.