Convert to LESS
I convert the styles.css to LESS and split it up to several files for
better overview.  I install the lesscss-resources plugin to handle these
files and compile them to CSS on the fly.  Because of the file splitting
I change the resource declarations to load only the files needed.