Modified closing balance issues
First, I implement a getter to obtain the modified closing balance of
invoices, dunnings, and credit memos.  The unify the computed value I
round it to the number of fraction digits specified in the application
configuration.  For it, I add all functions that return important
application configuration settings, such as the locale, the currency
symbol, the number of fraction digits etc., to the UserService class.
Furthermore, I refactor the code in ViewFilters to use the new methods
of UserService.

At last, I access the modifiedClosingBalance from the views and also
rename the attribute "data-closing-balance" to the more appropriate name
"data-modified-closing-balance" and refactor the CoffeeScript code as
well.