Closing balance issues
First, I add an additional field to the show views of invoices,
dunnings, and credit memos which display the closing balance.  For it, I
also add the invoicing-transaction.less file to these show views.

Then I implement a way to obtain the closing balance of invoices and
dunnings via AJAX.  This is necessary for credit memo forms which must
re-compute the modified closing balance in case the associated invoice
or dunning changes.

At last, I write a comprehensive documentation about the balancing topic
and suppress the pointer cursor on still unpaid links if they are in
state "paid too much".