Bugfixes in functional test cases
I fix some bugs in functional test cases which occurred after changing
common parts of the software for other the latest tests.

I also provide a fix for bug #268 which here appears when updating sales
items with pricing.  I prevent registering an LRU item if the domain
model instance contains errors.  I think the solution is not the best
but currently, it works.