Markdown in textareas
I start implementing the usage of Markdown in textareas, i. e. in
<textarea> fields.  I use the "markdown" Grails plugin which offers most
of the required features.  I implement the conversion of Markdown to
HTML in show views.

Furthermore, I improve the look in invoicing transaction show views and
add the missing resource declaration in theses views.

Also, I upgrade the Quartz plugin.