Simplify ticket structure
First, I simplify the ticket structure by removing the redundant
TicketMessage class and integrating it into TicketLogEntry.  This is
because TicketMessage is a subset of TicketLogEntry.

Furthermore, I modularize the display of ticket log entries by providing
templates for each log entry action.  This enables a future development
of AJAX calls that return the HTML code of a log entry.

At last, I add handling of an empty helpdesk list.  If so, the user
either gets the possibility to create a helpdesk or she should contact
her administrator depending on the available permissions.