Bugfix #273
I fix bug #273 (printing no longer working).  There was a problem while
retrieving the User.helpdesks collection when the Hibernate session is
already closed.  Because we only need this collection to fullfil the m:n
relation structure I implement a domain model class for the relation
table and change the way the users are added to the helpdesk.