Address in ticket; bug fixes
First, I change the street, location etc. properties in Ticket to an
embedded reference to class Address and change the views and database
structure.

Second, I fix a bug in print version.  After using Address objects in
various classes the generated XML has changed and thus, had to be
adapted.