Script to generate standalone version
I implement a script that calls the build-standalone target and then
adds some Windows batch files and README files in order to be able to
run SpringCRM on Windows.  All collected files are packaged into a zip
file for deployment.

Furthermore, I revise the log4j settings and remove the unneeded "live"
environment.