Implement issue #20
I work on issue #20 and implement a separate number of fraction digits
for prices which are used externally.  I add a new configuration setting
on the currency page within the system settings area.  Then I apply the
new setting to all currency values in the application.

Furthermore, I create a new JavaScript object $I which contains the i10n
settings such as decimal and grouping separator as well as both the
number of fraction digits.

I also fix a bug that occurred when clicking the "still unpaid" link.
That link had to be clicked more than once in order to set the value in
the payment amount field.