Attempt to export deep XML for FO
I add a XML marshaller and register it in the Bootstrap.groovy file.
The marshaller converts a given domain object deeply, that is, including
all referred objects.  Furthermore, it converts properties marked by the
XmlExport annotation.

However, this approach is very expensive.  Due to the deep conversion it
also converts the organization associated to an invoicing transaction as
well as all its subordinary objects, such as persons, other invoicing
transactions, calls, calendar entries etc.  In a worst case this can be
very time- and memory-consuming.