Issue #24
I implement issue #24: configuration of service for dunning charge and
default interest.  I add a configuration setting to the "numbering"
page of the system settings using an autocompleteex field.

Furthermore, I add support for functions in the option "valueInput" to
the autocompleteex widget.  Doing so, the caller may specify a function
which obtains the associated value input control.

Also, I add method "removeConfig" to class ConfigHolder to remove
configuration settings.