Issue #26
I implement issue #26: textarea for invoicing transaction subject.  I
implement a new facet of textarea fields which may be controlled by the
attribute "nl2br".  This displays a textarea as normal in the form and
uses the <g:nl2br> tag in the show view.  Furthermore, I change the
print template.