Create discrete mail service
I create a discrete service class to send mails in a general fashion:
MailSystemService.  Even if not implemented yet, this class should be
used to send all mails from the system because it will use the mail
configuration from the system settings.

I also update the TicketService class and the corresponding test
specifications to use the new mail service.