Download and permissions in document list
First, I implement a download feature when the user clicks a file in the
document list.  The download is realized using a hidden <iframe>.

Then, I add a further column to the document list containing the read
and write permissions of the file or folder.

At last, I write unit tests and CoffeeScript documentation.