Change on remove handling in document list
I remove the code that clients of the document list widget can specify a
callback that is invoked with a Promise object in order to add handler
functions.  This is much too complicate, why I introduce two new
callbacks: removeSuccess and removeFailed.

Furthermore, I add a new QUnit test what happens when a user wants to
remove an item but an AJAX failure occurred.