Set scale for fix point number values
Define a scale for each fix point number (BigDecimal) value and change
the database migration file according to it.  Also add non-null
constraints to database migration.  Set the maximum number of fraction
digits for currency values to 6 in system configuration.

Fix a bug that the hint to select panels on overview page was visible
even if panels were displayed.