Issue #47: sales journal not visible in demo
Fix issue #47 that the sales journal and other reports are not visible
in demo site.  The bug occurred because a user without the permission
to call the report controller was able to select the appropriate
navigation item.

First, convert the former class Modules to an enumeration Module and
move it to src/groovy.  Migrate all calls to the former Modules methods
to more suitable method calls to Module.  Change all module names to
upper underline style.  Add a database migration script that converts
the allowed modules in user_data table.

Fix a bug that allowed modules were stored as list of lists in User
object.