Issue #92
Fix an issue that occurred when invoices or reminders have been paid
completely.  The issue caused that invoices or reminders were still displayed
as unpaid, sometimes with negative balance.

Fixed by rounding of balance to number of digits used for external
representation because the balance and the payment amount is always specified
with that precision.