Settings for list of unpaid bills
Implement a dialog which allows to set options for the list of unpaid bills:

- minimum amount to pay
- sort and order criterion
- number of items in the list

For it, implement a generic way to specify additional HTML code in the header
of a panel.  Currently, this is visible after the first reload of the overview
page, only.  This is not a bug -- this is a feature.

Also fix an issue that the empty list message was not displayed when the last
panel had been closed.