Fix Hibernate issues
Fix an issue which occurs when saving multiple user settings at once.  This was
caused by a binding of a UserSettings object to the Credential object where the
latter was stored in the HTTP session.  The UserSettings object, in turn,
referenced a User object and one or more UserSetting objects.  This, in fact,
rendered the plan to not hold User objects in the HTTP session void.

The actual bug occurred when saving multiple UserSetting objects after another.
The first call to UserSetting.save() detects that the associated User object is
not part of the Hibernate session.  Thus, it loads the User object.  After
that, the User.afterLoad() method is called which in turn initializes a new
UserSettings object with a new set of UserSetting objects.  These new
UserSetting objects are now part of the Hibernate session.  When the second
UserSetting object was about to be saved, the error occurred.  This was because
the UserSetting object to save has the same ID as the UserSetting object in the
Hibernate session but they are different objects.